Created relations between several tables addWorkout methods were created for UserService and PlanService, also typo fixed

+++ src/main/java/com/example/traininapp/UserPack/UserRepo.java
@@ -7,8 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepo extends JpaRepository<User, Integer> {
+public interface UserRepo extends JpaRepository<User, Long> {
     @Query
     Optional<User> findByEmail(String email);
-
 }
